104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	unsigned long int a = 0, b = 1, c;
+	unsigned long long a = 0, b = 1, c;
 
 	int i;
